requirements.txt
Pillow==8.1.0
streamlit==0.78.0
Keras==2.4.3
opencv-contrib-python==4.5.1.48
